requirements.txt
alembic==1.8.1
anyio==3.6.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Mako==1.2.1
MarkupSafe==2.1.1
orjson==3.7.11
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.2
